+++ iqm-pulse/src/iqm/pulse/gates/rz.py
@@ -251,7 +251,7 @@ class RZ_ACStarkShift_smoothConstant(  # type: ignore[call-arg]  # type: ignore[
 class RZ_PRX_Composite(CompositeGate):
     """RZ gate implemented as a sequence of PRX gates."""
 
-    registered_gates = ["prx"]
+    registered_gates = ("prx",)
 
     def __init__(self, parent, name, locus, calibration_data, builder):
         super().__init__(parent, name, locus, calibration_data, builder)